requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchxrayvision>=1.0.0
grad-cam>=1.4.8
gradio>=4.0.0
pydicom>=2.4.0
Pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-image>=0.21.0
opencv-python>=4.8.0
pandas>=2.0.0
plotly>=5.15.0
fpdf2>=2.7.0
reportlab>=4.0.0
scipy>=1.11.0
seaborn>=0.12.0
